Amélioration du service Upload

+++ src/Service/Upload.php
@@ -4,26 +4,32 @@ namespace App\Service;
 
 class Upload
 {
-    public function upload($directory, $file, $oldFile = '')
+    public function upload($directory, $file, $inputError, $oldFile = '')
     {
-        if($file->guessExtension() == 'jpg' || $file->guessExtension() == 'jpeg' || $file->guessExtension() == 'png')
+        if(isset($file))
         {
-            if ($oldFile != '')
+            if($file->guessExtension() == 'jpg' || $file->guessExtension() == 'jpeg' || $file->guessExtension() == 'png')
             {
-                unlink($directory .'/'. $oldFile); 
-            }
-            $fileName = md5(uniqid()) . '.' . $file->guessExtension();
+                if ($oldFile != '')
+                {
+                    unlink($directory .'/'. $oldFile); 
+                }
+                $fileName = md5(uniqid()) . '.' . $file->guessExtension();
 
-            $file->move(
-                $directory,
-                $fileName
-            );
-    
-            return $fileName;
-        }
-        else
-        {
-            return false;
+                $file->move(
+                    $directory,
+                    $fileName
+                );
+                
+
+                return $fileName;
+                
+            }
+            else
+            {
+                $inputError->addError(new FormError("Le format d'image n'est pas accepté (jpg, jpeg, png)."));
+                return false;
+            }
         }
     }
 }
